Added check for if no databse exists. Converted hello world to database call.

+++ resetDB.py
@@ -1,6 +1,11 @@
 import app
 import os
 
-os.remove("test.db")
+if os.path.exists("test.db"): 
+    os.remove("test.db")
 app.db.create_all()
+
+app.db.session.add(app.User(username="Eugene"))
+app.db.session.add(app.Daily(name="Check Blood Sugar", user = 1))
+app.db.session.commit()
